+++ sideline/app/qa/onboarding/logger/page.tsx
@@ -1,0 +1,5 @@
+import { OnboardingQaLogger } from "@/components/qa/onboarding/OnboardingQaLogger";
+
+export default function QaOnboardingLoggerPage() {
+  return <OnboardingQaLogger />;
+}