+++ sideline/app/qa/onboarding/new-play-sheet/page.tsx
@@ -1,0 +1,5 @@
+import { OnboardingQaNewPlaySheet } from "@/components/qa/onboarding/OnboardingQaNewPlaySheet";
+
+export default function QaOnboardingNewPlaySheetPage() {
+  return <OnboardingQaNewPlaySheet />;
+}