+++ sideline/app/qa/onboarding/yardage/page.tsx
@@ -1,0 +1,5 @@
+import { OnboardingQaYardage } from "@/components/qa/onboarding/OnboardingQaYardage";
+
+export default function QaOnboardingYardagePage() {
+  return <OnboardingQaYardage />;
+}