+++ sideline/components/qa/onboarding/OnboardingQaBreakdown.tsx
@@ -1,0 +1,16 @@
+"use client";
+
+import { GuidedFirstDriveInsight } from "@/components/film/GuidedFirstDriveInsight";
+import { onboardingQaFirstDriveReadout } from "@/lib/onboardingQaFixture";
+
+export function OnboardingQaBreakdown() {
+  return (
+    <GuidedFirstDriveInsight
+      open
+      readout={onboardingQaFirstDriveReadout}
+      onCallAnotherDrive={() => {}}
+      onGoToFilmRoom={() => {}}
+      anotherDriveBusy={false}
+    />
+  );
+}